Move clang_cc to bpf_clang_cc
The reason for this package is that eBPF programs, as found in lo2s,
require clang regardless of the other compiler choice for compiling eBPF
programs. This allows to have clang in the store for lo2s specifically
and nothing else.

However the name `clang_cc` conflicts with the `clang_cc` provides
definition for the general-purpose clang_cc package.

Move it to the more specific bpf_clang_cc namespace.